+++ src/lib/browser/BrowserController.js
@@ -6,14 +6,9 @@ import Cryptography from '../Crypto'
 import packageJson from '../../../package.json'
 import BrowserAccountStorage from './BrowserAccountStorage'
 import uniqBy from 'lodash/uniqBy'
-
-import PQueue from 'p-queue'
 import Account from '../Account'
+import { STATUS_ALLGOOD, STATUS_DISABLED, STATUS_ERROR, STATUS_SYNCING } from '../interfaces/Controller'
 
-const STATUS_ERROR = Symbol('error')
-const STATUS_SYNCING = Symbol('syncing')
-const STATUS_ALLGOOD = Symbol('allgood')
-const STATUS_DISABLED = Symbol('disabled')
 const INACTIVITY_TIMEOUT = 7 * 1000
 const DEFAULT_SYNC_INTERVAL = 15
 
@@ -24,26 +19,29 @@ class AlarmManager {
 
   async checkSync() {
     const accounts = await BrowserAccountStorage.getAllAccounts()
+    const promises = []
     for (let accountId of accounts) {
       const account = await Account.get(accountId)
       const data = account.getData()
       const lastSync = data.lastSync || 0
       const interval = data.syncInterval || DEFAULT_SYNC_INTERVAL
+      if (data.scheduled) {
+        promises.push(this.ctl.scheduleSync(accountId))
+      }
       if (
         Date.now() >
         interval * 1000 * 60 + lastSync
       ) {
         // noinspection ES6MissingAwait
-        this.ctl.scheduleSync(accountId)
+        promises.push(this.ctl.scheduleSync(accountId))
       }
     }
+    await Promise.all(promises)
   }
 }
 
 export default class BrowserController {
   constructor() {
-    this.jobs = new PQueue({ concurrency: 1 })
-    this.waiting = {}
     this.schedule = {}
     this.listeners = []
 
@@ -71,8 +69,8 @@ export default class BrowserController {
     // Set up the alarms
 
     browser.alarms.create('checkSync', { periodInMinutes: 1 })
-    browser.alarms.onAlarm.addListener(alarm => {
-      this.alarms[alarm.name]()
+    browser.alarms.onAlarm.addListener(async alarm => {
+      await this.alarms[alarm.name]()
     })
 
     // lock accounts when locking is enabled
@@ -109,7 +107,7 @@ export default class BrowserController {
     // Setup service worker messaging
 
     // eslint-disable-next-line no-undef
-    if (typeof WorkerGlobalScope !== 'undefined' && self instanceof WorkerGlobalScope) {
+    if (!navigator.userAgent.includes('Firefox') && typeof WorkerGlobalScope !== 'undefined' && self instanceof WorkerGlobalScope) {
       addEventListener('message', (event) => this._receiveEvent(event.data, (data) => event.source.postMessage(data)))
     } else {
       browser.runtime.onMessage.addListener((data) => void (this._receiveEvent(data, (data) => browser.runtime.sendMessage(data))))
@@ -240,7 +238,6 @@ export default class BrowserController {
   }
 
   async scheduleSync(accountId, wait) {
-    console.log('called scheduleSync')
     if (wait) {
       if (this.schedule[accountId]) {
         clearTimeout(this.schedule[accountId])
@@ -260,27 +257,31 @@ export default class BrowserController {
       console.log('Account is already syncing. Not syncing again.')
       return
     }
-    if (!account.getData().enabled) {
+    // if the account is already scheduled, don't prevent it, to avoid getting stuck
+    if (!account.getData().enabled && !account.getData().scheduled) {
       console.log('Account is not enabled. Not syncing.')
       return
     }
 
-    if (this.waiting[accountId]) {
-      console.log('Account is already queued to be synced')
+    const status = await this.getStatus()
+    if (status === STATUS_SYNCING) {
+      await account.setData({ ...account.getData(), scheduled: account.getData().scheduled || true })
       return
     }
 
-    this.waiting[accountId] = true
-    await account.setData({ ...account.getData(), scheduled: true })
-
-    return this.jobs.add(() => this.syncAccount(accountId))
+    if (account.getData().scheduled === true) {
+      await this.syncAccount(accountId)
+    } else {
+      await this.syncAccount(accountId, account.getData().scheduled)
+    }
   }
 
   async scheduleAll() {
     const accounts = await Account.getAllAccounts()
     for (const account of accounts) {
-      this.scheduleSync(account.id)
+      await account.setData({...account.getData(), scheduled: true})
     }
+    this.updateStatus()
   }
 
   async cancelSync(accountId, keepEnabled) {
@@ -294,23 +295,24 @@ export default class BrowserController {
 
   async syncAccount(accountId, strategy) {
     console.log('Called syncAccount ', accountId)
-    this.waiting[accountId] = false
     if (!this.enabled) {
       console.log('Flocccus controller is not enabled. Not syncing.')
       return
     }
     let account = await Account.get(accountId)
-    await account.setData({ ...account.getData(), scheduled: false })
     if (account.getData().syncing) {
       console.log('Account is already syncing. Not triggering another sync.')
       return
     }
+    // executes long-running async work without letting the service worker to die
+    const interval = setInterval(() => browser.tabs.getCurrent(), 2e4)
     setTimeout(() => this.updateStatus(), 500)
     try {
       await account.sync(strategy)
     } catch (error) {
       console.error(error)
     }
+    clearInterval(interval)
     this.updateStatus()
   }
 
@@ -329,14 +331,14 @@ export default class BrowserController {
     }
   }
 
-  async updateBadge() {
+  async getStatus() {
     if (!this.unlocked) {
-      return this.setStatusBadge(STATUS_ERROR)
+      return STATUS_ERROR
     }
     const accounts = await Account.getAllAccounts()
     let overallStatus = accounts.reduce((status, account) => {
       const accData = account.getData()
-      if (status === STATUS_SYNCING || accData.syncing) {
+      if (status === STATUS_SYNCING || accData.syncing || account.syncing) {
         return STATUS_SYNCING
       } else if (status === STATUS_ERROR || (accData.error && !accData.syncing)) {
         return STATUS_ERROR
@@ -351,7 +353,11 @@ export default class BrowserController {
       }
     }
 
-    this.setStatusBadge(overallStatus)
+    return overallStatus
+  }
+
+  async updateBadge() {
+    await this.setStatusBadge(await this.getStatus())
   }
 
   async setStatusBadge(status) {
@@ -363,7 +369,11 @@ export default class BrowserController {
     }
 
     if (icon[status]) {
-      await browser.action.setIcon(icon[status])
+      if (navigator.userAgent.includes('Firefox')) {
+        await browser.browserAction.setIcon(icon[status])
+      } else {
+        await browser.action.setIcon(icon[status])
+      }
     }
   }
 
@@ -375,7 +385,6 @@ export default class BrowserController {
           await acc.setData({
             ...acc.getData(),
             syncing: false,
-            error: false,
           })
         }
       })