+++ src/ui/store/native/actions.js
@@ -4,7 +4,7 @@ import Logger from '../../../lib/Logger'
 import AdapterFactory from '../../../lib/AdapterFactory'
 import Controller from '../../../lib/Controller'
 import { i18n } from '../../../lib/native/I18n'
-import { CapacitorHttp as Http } from '@capacitor/core'
+import { Capacitor, CapacitorHttp as Http } from '@capacitor/core'
 import { Share } from '@capacitor/share'
 import Html from '../../../lib/serializers/Html'
 import { Bookmark, Folder } from '../../../lib/Tree'
@@ -137,7 +137,11 @@ export const actionsDefinition = {
   },
   async [actions.STORE_ACCOUNT]({ commit, dispatch, state }, { id,data }) {
     const account = await Account.get(id)
+    const oldData = account.getData()
     await account.setData(data)
+    if (oldData.localRoot !== data.localRoot) {
+      await account.init()
+    }
     commit(mutations.STORE_ACCOUNT_DATA, {id, data})
   },
   async [actions.TRIGGER_SYNC]({ commit, dispatch, state }, accountId) {
@@ -196,14 +200,18 @@ export const actionsDefinition = {
       throw new Error(i18n.getMessage('LabelLoginFlowError'))
     }
     let json = res.data
-    const browserWindow = await window.open(json.login, '_blank', 'toolbar=no,presentationstyle=pagesheet')
+    // iOS browser doesn't allow 3rd party cookies, so we have to open a standalone browser
+    const target = Capacitor.getPlatform() === 'ios' ? '_system' : '_blank'
+    const browserWindow = await window.open(json.login, target, 'toolbar=no,presentationstyle=pagesheet')
     do {
       await new Promise(resolve => setTimeout(resolve, 1000))
       try {
+        const data = new URLSearchParams()
+        data.set('token', json.poll.token)
         res = await Http.request({
           url: json.poll.endpoint,
           method: 'POST',
-          data: {token: json.poll.token},
+          data: data.toString(),
           headers: {'Content-type': 'application/x-www-form-urlencoded'}
         })
       } catch (e) {